requirements.txt
Flask>=2.3.0
tensorflow>=2.16.0
numpy>=1.24.0
Pillow>=10.0.0
Werkzeug>=2.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
xgboost>=1.7.0
gunicorn>=21.0.0